requirements.txt
flask==2.3.2
flask-cors==3.0.10
nltk==3.8.1
scikit-learn==1.2.2
tensorflow==2.12.0
pandas==1.5.3
pillow==9.5.0
opencv-python-headless==4.7.0.72
flask-wtf==1.1.1
flask-login==0.6.2
flask-babel==3.1.0